feat: move to core package
