+++ core/src/main/java/dev/derock/svcmusic/core/api/CommandSource.java
@@ -1,0 +1,5 @@
+package dev.derock.svcmusic.core.api;
+
+public interface CommandSource {
+
+}
