+++ core/src/main/java/dev/derock/svcmusic/core/api/MinecraftServer.java
@@ -1,4 +1,4 @@
-package dev.derock.svcmusic.api;
+package dev.derock.svcmusic.core.api;
 
 import java.util.List;
