+++ core/src/main/java/dev/derock/svcmusic/core/commands/NowPlayingCommand.java
@@ -1,15 +1,14 @@
-package dev.derock.svcmusic.commands;
+package dev.derock.svcmusic.core.commands;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
-import com.sedmelluq.discord.lavaplayer.track.AudioTrackInfo;
-import dev.derock.svcmusic.SimpleVoiceChatMusic;
-import dev.derock.svcmusic.audio.GroupManager;
-import dev.derock.svcmusic.audio.GroupSettingsManager;
-import dev.derock.svcmusic.audio.MusicManager;
-import dev.derock.svcmusic.util.ModUtils;
+import dev.derock.svcmusic.core.SimpleVoiceChatMusic;
+import dev.derock.svcmusic.core.audio.GroupManager;
+import dev.derock.svcmusic.core.audio.GroupSettingsManager;
+import dev.derock.svcmusic.core.audio.MusicManager;
+import dev.derock.svcmusic.core.util.ModUtils;
 import net.minecraft.command.CommandRegistryAccess;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
@@ -17,7 +16,7 @@ import net.minecraft.text.ClickEvent;
 import net.minecraft.text.Style;
 import net.minecraft.text.Text;
 
-import static dev.derock.svcmusic.util.ModUtils.checkPlayerGroup;
+import static dev.derock.svcmusic.core.util.ModUtils.checkPlayerGroup;
 
 public class NowPlayingCommand {
     public static void register(CommandDispatcher<ServerCommandSource> dispatcher, CommandRegistryAccess commandRegistryAccess, CommandManager.RegistrationEnvironment registrationEnvironment) {
