+++ core/src/main/java/dev/derock/svcmusic/core/commands/PauseCommand.java
@@ -1,24 +1,24 @@
-package dev.derock.svcmusic.commands;
+package dev.derock.svcmusic.core.commands;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import dev.derock.svcmusic.SimpleVoiceChatMusic;
-import dev.derock.svcmusic.audio.GroupManager;
-import dev.derock.svcmusic.audio.MusicManager;
-import dev.derock.svcmusic.util.ModUtils;
+import dev.derock.svcmusic.core.SimpleVoiceChatMusic;
+import dev.derock.svcmusic.core.audio.GroupManager;
+import dev.derock.svcmusic.core.audio.MusicManager;
+import dev.derock.svcmusic.core.util.ModUtils;
 import net.minecraft.command.CommandRegistryAccess;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.text.Text;
 
-import static dev.derock.svcmusic.util.ModUtils.checkPlayerGroup;
+import static dev.derock.svcmusic.core.util.ModUtils.checkPlayerGroup;
 
-public class ResumeCommand {
+public class PauseCommand {
     public static void register(CommandDispatcher<ServerCommandSource> dispatcher, CommandRegistryAccess commandRegistryAccess, CommandManager.RegistrationEnvironment registrationEnvironment) {
         dispatcher.register(CommandManager.literal("music")
-            .then(CommandManager.literal("resume")
-                .executes(ResumeCommand::execute)));
+            .then(CommandManager.literal("pause")
+                .executes(PauseCommand::execute)));
     }
 
     public static int execute(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
@@ -27,8 +27,8 @@ public class ResumeCommand {
 
         SimpleVoiceChatMusic.SCHEDULED_EXECUTOR.execute(() -> {
             GroupManager gm = MusicManager.getInstance().getGroup(result.group(), result.player().getServer());
-            gm.broadcast(Text.literal("Playback resumed by " + result.source().getName()));
-            gm.getPlayer().setPaused(false);
+            gm.broadcast(Text.literal("Playback paused by " + result.source().getName()));
+            gm.getPlayer().setPaused(true);
         });
 
         return 0;
