+++ core/src/main/java/dev/derock/svcmusic/core/commands/ResumeCommand.java
@@ -1,37 +1,34 @@
-package dev.derock.svcmusic.commands;
+package dev.derock.svcmusic.core.commands;
 
 import com.mojang.brigadier.CommandDispatcher;
-import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import dev.derock.svcmusic.SimpleVoiceChatMusic;
-import dev.derock.svcmusic.audio.GroupManager;
-import dev.derock.svcmusic.audio.MusicManager;
-import dev.derock.svcmusic.util.ModUtils;
+import dev.derock.svcmusic.core.SimpleVoiceChatMusic;
+import dev.derock.svcmusic.core.audio.GroupManager;
+import dev.derock.svcmusic.core.audio.MusicManager;
+import dev.derock.svcmusic.core.util.ModUtils;
 import net.minecraft.command.CommandRegistryAccess;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.text.Text;
 
-import static dev.derock.svcmusic.util.ModUtils.checkPlayerGroup;
+import static dev.derock.svcmusic.core.util.ModUtils.checkPlayerGroup;
 
-public class VolumeCommand {
+public class ResumeCommand {
     public static void register(CommandDispatcher<ServerCommandSource> dispatcher, CommandRegistryAccess commandRegistryAccess, CommandManager.RegistrationEnvironment registrationEnvironment) {
         dispatcher.register(CommandManager.literal("music")
-            .then(CommandManager.literal("volume")
-                .then(CommandManager.argument("volume_percent", IntegerArgumentType.integer(0, 100))
-                    .executes(VolumeCommand::execute))));
+            .then(CommandManager.literal("resume")
+                .executes(ResumeCommand::execute)));
     }
 
     public static int execute(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
-        int volume = IntegerArgumentType.getInteger(context, "volume_percent");
         ModUtils.CheckPlayerGroup result = checkPlayerGroup(context);
         if (result == null) return 1;
 
         SimpleVoiceChatMusic.SCHEDULED_EXECUTOR.execute(() -> {
             GroupManager gm = MusicManager.getInstance().getGroup(result.group(), result.player().getServer());
-            gm.broadcast(Text.literal("Volume set to " + volume + "% by " + result.source().getName()));
-            gm.setVolume(volume);
+            gm.broadcast(Text.literal("Playback resumed by " + result.source().getName()));
+            gm.getPlayer().setPaused(false);
         });
 
         return 0;
